Adds "Try This Style" modal to explore page
Implements a new modal that appears when a user presses on an image in the explore page.

The modal allows users to try the style of the selected image, which then opens the relevant generate modal with the image pre-selected. This improves the user experience by providing a more direct path to style transfer.

Adds animated scaling effect when image is pressed.