Enables style transfer from project details
Allows users to navigate to the style transfer modal directly from the project details page and confirmation step.

This change adds a 'styletransfer' option to the restyle menu and handles navigation to the style transfer modal.  It also saves the generated image as an alternative generation of the original project, and navigates back to the original project page when generation is complete.